Assessment/Smart_content_creation/requirements.txt
streamlit>=1.32.0
langchain>=0.1.14
langchain-google-genai>=0.0.8
google-generativeai>=0.3.2
autogen>=0.2.20
python-dotenv>=1.0.1
requests>=2.31.0